top/bottom navigation ui improvements and additions
