+++ public/locales/en.json
@@ -14,6 +14,7 @@
   "averagedBalance": "Averaged balance",
   "balance": "Balance",
   "bonus": "Bonus",
+  "borrow": "Borrow",
   "cancel": "Cancel",
   "changeAccount": "Change Account",
   "chooseToken": "Choose Token",
@@ -44,7 +45,7 @@
   "exchangeRate": "Exchange Rate",
   "failedTitle": "Transaction Failed.",
   "fast": "Fast",
-  "fee": "Fee",
+  "fee": "Protocol fee",
   "from": "From",
   "gas": "Gas",
   "getWallet": "Get a wallet!",
@@ -163,6 +164,7 @@
   "txSpeedUp_swap": "Swap transaction was sped up. View it on Etherscan.",
   "txSpeedUp_tokenApproval": "Token approval transaction was sped up. View it on Etherscan.",
   "unavailable": "Unavailable",
+  "underconstruction": "Under construction 👷",
   "unnecessaryApprovalAskA": "As some tokens (eg. USDT) do not comply to ERC20 standards for `approve` function, when using limited approvals, you may have to reset the approvals to 0 before changing to another value. If gas usage is your concern, we recommend using unlimited approvals. Otherwise, we recommend using limited approvals for added security.",
   "unnecessaryApprovalAskQ": "Reset token approval to 0 when having partial approval beforehand.",
   "unstake": "Unstake",
