+++ src/components/ToolTip.scss
@@ -24,22 +24,10 @@
     /* Position the tooltip */
     position: absolute;
     z-index: 1;
-    bottom: 100%;
+    // bottom: 100%;
     left: 50%;
     transform: translateX(-50%);
     margin-bottom: 15px;
-  }
-
-  /* Tooltip arrow */
-  & > .tooltipText::after {
-    content: "";
-    position: absolute;
-    top: 100%;
-    left: 40%;
-    margin-left: -5px;
-    border-width: 10px;
-    border-style: solid;
-    border-color: $primary transparent transparent transparent;
   }
 
   &:hover > .tooltipText {
