--- src/components/Version.module.scss
@@ -1,13 +1,0 @@
-@import "../styles/theme";
-
-div.version {
-  font-size: 15px;
-  position: absolute;
-  bottom: 24px;
-  left: 56px;
-}
-
-a {
-  margin-left: 10px;
-  display: inline-block;
-}
